requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pdfplumber==0.10.3
sqlalchemy==2.0.23
python-multipart==0.0.6
pandas==2.1.3
pydantic==2.5.0
openpyxl==3.1.5
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
